dataset for training custom DL model
This dataset contains more than 4 lakh rows and 3 columns with 6 classes. We used this dataset to train our Deep learning model to determine the emotion of user from the text input the user provides to our Chatbot. It detects the emotion of user from his/her input prompt and classifies the emotion as one of the 6 categories/classes (anger, love, joy, surprise, sad, happy) along with the confidence score. Then we use this information to prepare the prompt for the LLM API.